+++ src/routes/index.svelte
@@ -1,7 +1,7 @@
 <script>
 	import Calender from '../components/Calender.svelte';
 	import Hero from '../components/Hero.svelte';
-	import HeroDown from '../components/heroDown.svelte';
+	import HeroDown from '../components/sectionWrapper.svelte';
 </script>
 
 <svelte:head>
@@ -15,7 +15,7 @@
 	href="#termine"
 	note="Referenzschule der TU München"
 	image="https://lsg.musin.de/homepage/images/header-images/schulhof_mini.jpg"
-	awards="{['https://lsg.musin.de/homepage/images/LOGOsorsmc_SCREEN_80mm_RGB_mini.jpg', 'https://lsg.musin.de/homepage/images/delf.jpg', 'https://lsg.musin.de/homepage/images/cae.png', 'https://lsg.musin.de/homepage/images/cils.png', 'https://lsg.musin.de/homepage/images/KombiLogo.png']}" />
+	awards="{['https://lsg.musin.de/homepage/images/LOGOsorsmc_SCREEN_80mm_RGB_mini.jpg', 'https://lsg.musin.de/homepage/images/delf.jpg', 'https://lsg.musin.de/homepage/images/cae.png', 'https://lsg.musin.de/homepage/images/cils.png']}" />
 <HeroDown
 	header="Aktuelle Termine"
 	subHeader="Alle demnächst anstehenden Termine des Louise-Schroeder-Gynasiums"
